# src/back/documents/Swagger Fabrica de Historias.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Fabrica de Historias",
    "version": "1.0.0",
    "description": "API para cadastro de usuários, login e cadastro de obras."
  },
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "Servidor local"
    }
  ],
  "paths": {
    "/usuarios": {
      "post": {
        "summary": "Cadastrar novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário cadastrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "example": 1
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/{id}": {
      "get": {
        "summary": "Buscar usuário por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        }
      }
    },
    "/usuarios/login": {
      "post": {
        "summary": "Logar um usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário logado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas"
          }
        }
      }
    },
    "/obras/cadastrar": {
      "post": {
        "summary": "Cadastrar uma nova obra",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObraRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Obra cadastrada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Obra"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Usuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "senha": {
            "type": "string"
          },
          "tipoUsuario": {
            "type": "string"
          }
        }
      },
      "UsuarioRequest": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "senha": {
            "type": "string"
          },
          "tipoUsuario": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "senha": {
            "type": "string"
          },
          "tipoUsuario": {
            "type": "string"
          }
        }
      },
      "ObraRequest": {
        "type": "object",
        "properties": {
          "titulo": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "data": {
            "type": "string"
          },
          "numeroPaginas": {
            "type": "integer"
          },
          "idAutor": {
            "type": "integer"
          },
          "livro": {
            "type": "string"
          }
        }
      },
      "Obra": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "titulo": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "data": {
            "type": "string"
          },
          "numeroPaginas": {
            "type": "integer"
          },
          "idAutor": {
            "type": "integer"
          },
          "livro": {
            "type": "string"
          }
        }
      }
    }
  }
}